960: Moved file and removed unused test
