+++ web-client/src/presenter/sequences/updatePetitionDetailsSequence.js
@@ -1,5 +1,5 @@
 import { clearAlertsAction } from '../actions/clearAlertsAction';
-import { computePetitionDatesAction } from '../actions/StartCaseInternal/computePetitionDatesAction';
+import { computePetitionDatesAction } from '../actions/computePetitionDatesAction';
 import { computeStatisticDatesAction } from '../actions/StartCaseInternal/computeStatisticDatesAction';
 import { navigateToCaseDetailCaseInformationActionFactory } from '../actions/navigateToCaseDetailCaseInformationActionFactory';
 import { setAlertErrorAction } from '../actions/setAlertErrorAction';
